requirements.txt
arrow                 1.3.0
cbor2                 5.6.2
luma.core             2.4.2
luma.oled             3.13.0
paho-mqtt             2.0.0
pillow                10.2.0
pip                   23.0.1
pyftdi                0.55.0
pyserial              3.5
python-dateutil       2.9.0.post0
pyusb                 1.2.1
RPi.GPIO              0.7.1
setuptools            66.1.1
six                   1.16.0
smbus2                0.4.3
spidev                3.6
times                 0.7
types-python-dateutil 2.8.19.20240311
